+++ src/main/java/GradebookRepository.java
@@ -2,6 +2,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GradebookRepository extends JpaRepository<Gradebook, Long> {
+public interface GradebookRepository extends JpaRepository<Gradebook, Integer> {
 
 }
